+++ app-backend/batch/src/main/scala/stacExport/v2/CampaignExport.scala
@@ -2,6 +2,7 @@ package com.rasterfoundry.batch.stacExport.v2
 
 import com.rasterfoundry.common.S3
 import com.rasterfoundry.database.{
+  AnnotationLabelClassGroupDao,
   AnnotationLabelDao,
   AnnotationProjectDao,
   CampaignDao,
@@ -10,8 +11,8 @@ import com.rasterfoundry.database.{
 }
 import com.rasterfoundry.datamodel.TileLayerType.{MVT, TMS}
 import com.rasterfoundry.datamodel.{
+  AnnotationLabelClassGroup,
   AnnotationProject,
-  Scene,
   StacExport,
   TileLayer,
   UnionedGeomExtent
@@ -29,7 +30,7 @@ import com.azavea.stac4s.syntax._
 import com.google.common.util.concurrent.ThreadFactoryBuilder
 import doobie.Transactor
 import doobie.implicits._
-import geotrellis.vector.{Extent, MultiPolygon, Projected}
+import geotrellis.vector.Extent
 import io.circe.optics.JsonPath._
 import io.circe.syntax._
 import io.circe.{Encoder, Json}
@@ -82,6 +83,7 @@ object optics {
 
 case class ExportState(
     exportDefinition: StacExport,
+    labelGroupOpt: Option[List[AnnotationLabelClassGroup]],
     rootCatalog: StacCatalog,
     remainingAnnotationProjects: List[AnnotationProject],
     annotationProjectImageryItems: Map[
@@ -437,11 +439,14 @@ class CampaignStacExport(
 
   def run(): IO[Option[ExportData]] =
     for {
-      campaign <- CampaignDao.getCampaignById(campaignId).transact(xa)
-      childProjects <- campaign traverse { campaign =>
+      campaignOpt <- CampaignDao.getCampaignById(campaignId).transact(xa)
+      labelGroupOpt <- campaignOpt traverse { campaign =>
+        AnnotationLabelClassGroupDao.listByCampaignId(campaign.id).transact(xa)
+      }
+      childProjects <- campaignOpt traverse { campaign =>
         AnnotationProjectDao.listByCampaign(campaign.id).transact(xa)
       }
-      initialStateO = (campaign, childProjects) mapN {
+      initialStateO = (campaignOpt, childProjects) mapN {
         case (campaign, projects) =>
           val rootCatalog = StacCatalog(
             stacVersion,
@@ -453,6 +458,7 @@ class CampaignStacExport(
           )
           ExportState(
             exportDefinition,
+            labelGroupOpt,
             rootCatalog,
             projects,
             Map.empty,
@@ -496,25 +502,6 @@ class CampaignStacExport(
   ): ExportState => ExportState =
     optics.labelAssetsLens.modify(_ ++ toAppend)
 
-  def imageryItemFromScene(
-      scene: Scene,
-      annotationProject: AnnotationProject
-  ): Option[newtypes.SceneItem] = {
-    val s3ImageLocation = scene.ingestLocation
-    val footprint = scene.dataFootprint
-    val sceneCreationTime =
-      scene.filterFields.acquisitionDate orElse Some(scene.createdAt)
-    (s3ImageLocation, footprint, sceneCreationTime) mapN {
-      case (url, footprint, timestamp) =>
-        makeSceneItem(
-          url,
-          footprint,
-          timestamp.toInstant,
-          annotationProject
-        )
-    }
-  }
-
   private def imageryItemFromTileLayers(
       annotationProject: AnnotationProject,
       taskStatuses: List[String]
@@ -541,18 +528,11 @@ class CampaignStacExport(
   ): IO[ExportState] = {
     for {
       // make the catalog for this annotation project
-      // make the scene item for this annotation project with an s3 asset
-      scene <- AnnotationProjectDao
-        .getFirstScene(annotationProject.id)
-        .transact(xa)
-      imageryItemO <- scene.fold(
-        imageryItemFromTileLayers(
-          annotationProject,
-          inputState.exportDefinition.taskStatuses
-        )
-      )({ scene =>
-        IO.pure(imageryItemFromScene(scene, annotationProject))
-      })
+      // make the scene item for this annotation project with a tile layer asset
+      imageryItemO <- imageryItemFromTileLayers(
+        annotationProject,
+        inputState.exportDefinition.taskStatuses
+      )
       imageryItemsAppend = imageryItemO map { (item: newtypes.SceneItem) =>
         Map(newtypes.AnnotationProjectId(annotationProject.id) -> item)
       } getOrElse Map.empty
@@ -560,7 +540,8 @@ class CampaignStacExport(
       featureGeoJSON <- AnnotationLabelDao
         .getAnnotationJsonByTaskStatus(
           annotationProject.id,
-          inputState.exportDefinition.taskStatuses
+          inputState.exportDefinition.taskStatuses,
+          inputState.labelGroupOpt
         )
         .transact(xa)
       labelAssetAppend = featureGeoJSON map { geojson =>
@@ -577,7 +558,10 @@ class CampaignStacExport(
         )
         .transact(xa)
       labelItemExtensionO <- AnnotationProjectDao
-        .getAnnotationProjectStacInfo(annotationProject.id)
+        .getAnnotationProjectStacInfo(
+          annotationProject.id,
+          inputState.labelGroupOpt
+        )
         .transact(xa)
       labelItemsAppend = (taskExtent, labelItemExtensionO, imageryItemO) mapN {
         case (extent, labelItemExtension, imageryItem) =>
@@ -640,43 +624,6 @@ class CampaignStacExport(
         ),
         Nil,
         assets,
-        None,
-        Map("datetime" -> createdAt.asJson).asJsonObject
-      )
-    )
-  }
-
-  private def makeSceneItem(
-      url: String,
-      footprint: Projected[MultiPolygon],
-      createdAt: Instant,
-      annotationProject: AnnotationProject
-  ): newtypes.SceneItem = {
-    val latLngGeom = footprint.withSRID(4326)
-    val latLngExtent = Extent(latLngGeom.geom.getEnvelopeInternal)
-    newtypes.SceneItem(
-      StacItem(
-        s"${UUID.randomUUID}",
-        stacVersion,
-        Nil,
-        "Feature",
-        footprint.withSRID(4326),
-        TwoDimBbox(
-          latLngExtent.xmin,
-          latLngExtent.ymin,
-          latLngExtent.xmax,
-          latLngExtent.ymax
-        ),
-        Nil,
-        Map(
-          "data" -> StacItemAsset(
-            url,
-            Some(s"COG for project ${annotationProject.name}"),
-            None,
-            Set(StacAssetRole.Data),
-            Some(`image/cog`)
-          )
-        ),
         None,
         Map("datetime" -> createdAt.asJson).asJsonObject
       )